move gob test to dedicated file
